fix(types): clean mypy, and a type check that was checking nothing
mypy reported 47 errors. Thirteen were the same mistake: a parameter
annotated X with a default of None, which PEP 484 forbids and mypy has
refused by default for years. The annotations were wrong, not the defaults.

The rest were real.

core/config.py defines Settings twice, once with pydantic-settings and once
without. That is deliberate and now says so instead of looking like an
accident, and its docstring pointed at a server extra that no longer exists.

training.py reassigned a str parameter to a torch.device under the same name,
so the string a caller passed and the resolved device object shared one name.
The resolved device has its own now, and the model, the data and
train_one_epoch all receive it.

benchmark.py returned numpy float64 from a function declared to return float.
That is not pedantry: a float64 serialises differently through json and
compares differently against a plain float.

Two reconstruction classes declared the attribute holding their result as
None and then assigned an array to it. pipelines/recovery.py called a model
attribute that is Optional, which is correct because load_model runs first,
and now records that order where a reader will see it.

Fourteen were no-any-return: a function annotated to return an ndarray, and
returning one, reported as returning Any because scipy's stubs say Any. That
is a statement about the stubs rather than about this code. The rule is
turned off in configuration, where the reason is written down, rather than
silenced on fourteen separate lines.

The worst finding was the gate itself. The lint job ran mypy on Python 3.10
with only ruff and mypy installed, so it type checked against no numpy, no
torch and no scipy. Every array operation was Any, and a clean result meant
nothing at all. It now installs the package with its extras first.

It also runs on 3.12 now. numpy ships stubs written in PEP 695 syntax that
mypy refuses to parse when told to target an older Python, so the run failed
on numpy rather than on this code. The test matrix still covers 3.10 through
3.12, and the matrix job keeps mypy as reporting only for the same reason.
The command line no longer passes --ignore-missing-imports, which was
overriding the configured settings.

mypy: no issues in 45 source files. ruff clean, formatter clean, 26 tests
pass, and the end to end reconstruction is unchanged at 4.479 to 26.436 dB.

+++ src/promethium_seismic/ml/benchmark.py
@@ -96,11 +96,14 @@ class BenchmarkEngine:
                 ssim_list.append(ssim_val)
 
         metrics = {
-            "mse": np.mean(mse_list),
-            "mae": np.mean(mae_list),
-            "snr": np.mean(snr_list),
-            "psnr": np.mean(psnr_list),
-            "ssim": np.mean(ssim_list),
+            "mse": float(np.mean(mse_list)),
+            # float() rather than the numpy scalar: the return type says
+            # float, and a float64 serialises differently through json and
+            # compares differently against a plain float.
+            "mae": float(np.mean(mae_list)),
+            "snr": float(np.mean(snr_list)),
+            "psnr": float(np.mean(psnr_list)),
+            "ssim": float(np.mean(ssim_list)),
         }
 
         logger.info(f"Benchmark Results: {json.dumps(metrics, indent=2)}")
